requirements.txt
selenium
requests
mysql-connector-python
beautifulsoup4
